continued

+++ trektech/ch/awae/trektech/Properties.java
@@ -2,6 +2,6 @@ package ch.awae.trektech;
 
 public class Properties {
 
-	public static final short PLASMA_TRANSFER_SPEED = 100;
-	
+	public static final int PLASMA_TRANSFER_SPEED = 100;
+
 }
